Delete legacy KeychainService class; complete the migration
Final cleanup that removes the last KeychainService.shared reference in
app code and deletes VivaDicta/Services/KeychainService.swift entirely.

Three apiKey accessors that previously lived in multi-target files
(KeychainService.swift, CloudModel.swift, TranscriptionModelProvider.swift)
move into main-target-only sibling extension files so the multi-target
parents (CloudModel, TranscriptionModelProvider) compile cleanly in
the share, action, and keyboard extensions which don't link Keychain:

- AIProvider.apiKey -> VivaDicta/Services/AIProvider+APIKey.swift
  (uses a file-private KeychainServiceImpl(); domain enum used by
  non-DI types like data models, so an injected dependency would
  balloon scope)
- CloudModel.apiKey -> VivaDicta/Models/CloudModel+APIKey.swift
- TranscriptionModelProvider.discoverCloudProvider() ->
  VivaDicta/Models/TranscriptionModelProvider+Discovery.swift

pbxproj cleanup removes the three Services/KeychainService.swift entries
from extension membershipExceptions lists.

Verified: full app test suite passes (323/323); zero
KeychainService.shared references remain anywhere in app, extensions,
or tests; the KeychainService class no longer exists.
